feat(fontconfig): add Helvetica to Arial alias
In some PDFs, the authors failed to embed the fonts they used. This
makes some PDF readers use very strange font substitutions when they try
and `fc-match` for a reasonable font. This specifically aliases
Helvetica to Arial, which is supposedly a good match [1]:

> Microsoft commissioned a copy of Helvetica – The well-known
> Arial. (ttf-mscorefonts-installer)

So, to use this, `ttf-mscorefonts-installer` must also be installed. The
alias itself and location was based on [2].

[1]: https://askubuntu.com/questions/599915/what-is-the-closest-font-to-helvetica-available-on-ubuntu
[2]: https://jichu4n.com/posts/how-to-set-default-fonts-and-font-aliases-on-linux/